frontend, docker: watch for new changes and use JS from the static mount
This is useful in general. But we also have an issue that when both
frontend and backend containers are started for the first time, it
takes much longer for the frontend to be up-and-running and to compile
JavaScript output. Once it is done, backend doesn't notice it and
therefore frontend code doesn't work. Backend needs to be
restarted. This commit fixes the issue.