frontend, backend: first draft of an AI prompt
I am adding a first version of the Log Detective service prompt. There
is not a link on the website yet because the feature isn't finished
You can access it via http://127.0.0.1:5020/explain

The frontend is fully-functional, few things are remaining:

- It needs to be finished from a visual standpoint
- The input form uses only a simple textarea, the integration with
  Copr, Koji, etc that we have on our homepage needs to be added

The backend part is only a mock implementation, always returning the
same hardcoded values for every input. Somebody needs to implement an
actual communication with the Log Detective service.