Connect to server and provide data to frontend
Connect website with the logdetective server. This code will connect to
the logdetective server instance and provide it URL to logs. The URL
is taken from environment variable.

After getting response from the server it will process it.

Also downloads the original log content so user is able to view it
easily without switching tabs.