frontend: fix missing text between snippets
The previous code printed:

1. Text from the start of the log to the first snippet
2. Take two snippets
   2.1. Text of the first snippet
   2.2. Text between first and second snippet
   2.3 Text of the second snippet
3. Repeat until there are no snippets left
4. Text after the last snippet until the end of the file

This works great for two snippets but notice the bug when there is more of
them. Let's say we have 6 snippets, then there is going to be a missing text
betwen 2 and 3rd snippet, 4th and 5th snippet, and finally 6th and 7th snippet.

This commit fixes that.