Fix broken review page
The new fields to `OkResponse` introduced in #260 cause an error (#265).

This tries to fix it by allowing `OkResponse.review_id` to be either a
`str` or an `uuid.UUID`. This is for backwards compatibility when
reviews were just a timestamp. Remember, we need to handle those when
reviewing older contributions.

The `/frontend/review` endpoint now returns a properly filled
`OkRepsonse`. This part is the actual fix for #265.